refactor: reorganize SignUp class and package
- Moved SignUp class to a new package.
- Updated references in the manifest and other classes.
- Introduced a Popup utility for displaying messages.
