+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.PathPilotandroidclient"
         tools:targetApi="31">
         <activity
-            android:name=".SignUp"
+            android:name=".signup.SignUp"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
